+++ CHANGELOG.md
@@ -13,6 +13,9 @@ IMPROVEMENTS:
 
 BUG FIXES:
 
+- mobile: Use fork of badgerdb in mobile. Fixes the directory lock issue with
+          Android 6.
+
 ## v0.6.2 (February 7, 2020)
 
 IMPROVEMENTS: